[graphic] moved message to better folder
