+++ CMakeLists.txt
@@ -25,6 +25,9 @@ include(cmake/Dependencies/OGRE.cmake)
 
 set(EXECUTABLE_OUTPUT_PATH "${CMAKE_BINARY_DIR}/bin/")
 
+# thanks Ogre 3d
+add_compile_options("-Wno-ignored-attributes")
+
 # ----------------------------
 # External dependencies
 # ----------------------------
