+++ src/libgraphic/src/message/entity/AnimateGraphicEntityMessage.cpp
