+++ src/libgraphic/src/message/entity/AnimateGraphicEntityMessage.h
