+++ src/libgraphic/src/message/entity/MoveGraphicEntityMessage.cpp
