+++ src/libgraphic/src/message/entity/NewGraphicEntityMessage.cpp
