+++ src/libgraphic/src/message/entity/NewGraphicEntityMessage.h
